refactor manager and after submit, selectedBook reset
